requirements.txt
asgiref==3.10.0
channels==4.3.1
channels_redis==4.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.2
pillow==12.0.0
PyJWT==2.10.1
redis==7.0.1
sqlparse==0.5.3
